rename enum and interfaces, change type of name in class Animal
