+++ src/Animals/SmallDog.ts
@@ -1,8 +1,8 @@
-import { AnimalNames } from '../Enums/enumAnimalNames';
-import { IProtectHerd } from '../Interfaces/IProtectHerd';
+import { AnimalNames } from '../Enums/AnimalNamesEnum';
+import { ProtectHerdInterface } from '../Interfaces/ProtectHerdInterface';
 import { Animal } from './Animal';
 
-export class SmallDog extends Animal implements IProtectHerd {
+export class SmallDog extends Animal implements ProtectHerdInterface {
   constructor() {
     super(AnimalNames.SMALL_DOG, 6);
   }
