+++ src/Animals/Wolf.ts
@@ -1,14 +1,14 @@
-import { AnimalNames } from '../Enums/enumAnimalNames';
-import { IClearHerd } from '../Interfaces/IClearHerd';
+import { AnimalNames } from '../Enums/AnimalNamesEnum';
+import { AttackHerdInterface } from '../Interfaces/AttackHerdInterface';
 import { Animal } from './Animal';
 
-export class Wolf extends Animal implements IClearHerd {
+export class Wolf extends Animal implements AttackHerdInterface {
   constructor() {
     super(AnimalNames.WOLF);
   }
 
   // TODO: implement when herd is ready
-  clearHerd(herd: Herd): boolean {
+  attackHerd(herd: Herd): boolean {
     return true;
   }
 }
