+++ src/Enums/AnimalNamesEnum.ts
