+++ src/Interfaces/AttackHerdInterface.ts
@@ -1,0 +1,3 @@
+export interface AttackHerdInterface {
+  attackHerd(herd: Herd): boolean;
+}
