--- src/Interfaces/IClearHerd.ts
@@ -1,3 +1,0 @@
-export interface IClearHerd {
-  clearHerd(herd: Herd): boolean;
-}
