--- src/Interfaces/IProtectHerd.ts
@@ -1,3 +1,0 @@
-export interface IProtectHerd {
-  protectHerd(herd: Herd): void;
-}
