+++ src/Interfaces/ProtectHerdInterface.ts
@@ -1,0 +1,3 @@
+export interface ProtectHerdInterface {
+  protectHerd(herd: Herd): void;
+}
